Fix/lab 415 (#1050)
* Update Forum link to include default category ID in URL

The Forum link now includes a query parameter to set a default category (cid=1). This ensures users are directed to the appropriate category upon navigation.

* Set default value for 'cid' in Feed component

Previously, 'cid' could be null if not provided in search parameters, potentially causing issues. This change ensures a default value of 1 is set, improving robustness and preventing unexpected behavior.