Fix/lab 470 (#1077)
* Add duplicate <ItemMenu> and 'use client' directive

Added an additional <ItemMenu> component in the Post component's UI, potentially duplicating the menu for posts. Also added the 'use client' directive to the ItemMenu component to ensure client-side rendering.

* Remove duplicate ItemMenu component rendering

The ItemMenu component was being rendered twice in the same section. This change removes the redundant instance to ensure correct UI behavior and avoid unnecessary duplication.